treewide: Restructure for unit tests
