+++ main.cpp
@@ -1,4 +1,4 @@
-#include "Checksec.h"
+#include "checksec.h"
 
 using namespace std;
