+++ test/winchecksec-test.cpp
@@ -1,0 +1,1 @@
+#include <checksec.h>
